[api][integration] Move mcp module from api to integration.
